Create VegaChart mixin
