+++ src/candela/components/ScatterPlot/index.js
@@ -1,8 +1,8 @@
 import VisComponent from '../../VisComponent';
-import vega from '../../util/vega';
+import VegaChart from '../../VisComponent/mixin/VegaChart';
 import spec from './spec.json';
 
-export default class ScatterPlot extends VisComponent {
+export default class ScatterPlot extends VegaChart(VisComponent, spec) {
   static get spec () {
     return {
       options: [
@@ -77,27 +77,5 @@ export default class ScatterPlot extends VisComponent {
         }
       ]
     };
-  }
-
-  constructor (el, options) {
-    super(el);
-    this.options = options;
-  }
-
-  render () {
-    this.chart = vega.parseChart(spec, this.el, this.options);
-  }
-
-  get serializationFormats () {
-    return ['png', 'svg'];
-  }
-
-  serialize (format) {
-    if (!this.chart) {
-      return Promise.reject('The render() method must be called before serialize().');
-    }
-    return this.chart.then(vobj => {
-      return vobj.toImageURL(format);
-    });
   }
 }
